+++ src/mcp-bundle/src/DependencyInjection/McpResourcePass.php
@@ -11,10 +11,10 @@
 
 namespace Symfony\AI\McpBundle\DependencyInjection;
 
-final class McpToolPass extends AbstractMcpPass
+final class McpResourcePass extends AbstractMcpPass
 {
     protected function getTag(): string
     {
-        return 'mcp.tool';
+        return 'mcp.resource';
     }
 }